Problem schema changes, new testproblem.
Added new fields to the schemas with function expressions. These fields
contain information on whether the expression is differentiable,
convex, or linear. Added also a method to the Problem model that can
infer the vairable domain (continuous, integer, or mixed) of the
problem. Added a simple knapsack test problem with three objectives.